requirements.txt
flask==3.0.3
twilio==9.3.4
anthropic==0.34.2
gspread==6.1.2
google-auth==2.34.0
gunicorn==22.0.0
